BlogAPI/requirements.txt
Flask
Flask-SQLAlchemy
Flask-JWT-Extended
Flask-Bcrypt
Flask-CORS
marshmallow
python-dotenv
pytest
pytest-flask
